Use this.get('currentURL') in updateScrollPosition
Previously we retrieved the `currentURL` from the `handler` object.  And
did so by finding the `handler`'s `router`.  Which was recently made
into an intimiate API until Ember 3.5.

To remove the deprecation we detect whether the underscored intimate API
is available and use it if it is.  Otherwise we defaulted back to using
the deprecated `router` api.

Fortunately, we know that this mixin is added to the `Router` (as per
the readme) and `currentURL` is public api on that object.  So within
our Mixin we can use `this` as the `Router` and bypass the need to
retreive it from the handler.

This change necessitated that we modify the way we mock our tests.
Previously we created a fake transition that handled how our URL needed
to behave.  In order to preserve the current test API I hoisted
`subject` up and added the `currentURL` to it directly when our
transition mock is executed.

It would be more correct for us to define the `currentURL` property on
the `RouterScrollObject` when we mixin in our RouterScroll mixin, but
modifying the instance seems just as valid for our testing purposes.