Fix scrolling to an element with an ID specified in the URL's hash segment (#302)
Ember router scroll supports automatically scrolling to an element with a given ID when the URL contains a hash segment

For instance, `/users/123#account` must scroll to `<div id="account">` automatically, or so was the theory.
The reason why this didn't work is that the existence of an element with that ID is checked when the router scroll in scheduled, not when it's performed. Given how common it is to have loading substates, it is very likely that by the time the scroll is scheduled the element we intend to scroll to is not yet in the page.

This PR moves the moment in which we check if the element exists to the actual moment when we scroll instead of the moment when we schedule it. That makes the feature work with loading substates.